feat: can delete pets now

+++ src/components/organisms/PetCard.jsx
@@ -3,56 +3,55 @@ import Image from 'next/image'
 import { redirect, useParams } from 'next/navigation';
 import { useRouter } from 'next/router'
 import { useContext } from 'react';
+import { useToast } from "@chakra-ui/react";
+
 export default function PetCard({pet}) {
-
   const router = useRouter();
-  const params = useParams();
-
+  const toast = useToast();
   const { token } = useContext(GlobalContext);
-
-  async function handleDelete() {
+  const url = process.env.NEXT_PUBLIC_API_URL
+  
+  async function handleDelete(pet_id) {
     
-    const response = await fetch(`${url}/pets/${params}`, {
-      method: "DELETE",
-      headers: {
-          Accept: "application/json",
-          Authorization: "Bearer " + token 
-      },
-    });
-
-    const result = await response.json();
-    console.log(result);
-    
-    if(result.errors) {
-        setErrors(result.errors)
-    }
-
-    if (!response.ok) {
-      toast({
-        title: 'Login failed',
-        description: 'Please try again...',
-        status: 'error',
-        duration: 3000,
-        isClosable: true
-      })
-    } else {
-      toast({
-        title: 'Successfully created pet!',
-        description: 'Redirecting you back...',
-        status: 'success',
-        duration: 3000,
-        isClosable: true
-      })
-      setTimeout(function () {
-        router.push('/profile')
-      }, 1000)
+    if(url && pet_id) {
+      const response = await fetch(`${url}/pets/${pet_id}`, {
+        method: "DELETE",
+        headers: {
+            Accept: "application/json",
+            Authorization: "Bearer " + token 
+        },
+      });
+  
+      const result = await response.json();
+      console.log(result);
+      
+      if (!response.ok) {
+        toast({
+          title: 'Failed to delete',
+          description: 'Please try again...',
+          status: 'error',
+          duration: 3000,
+          isClosable: true
+        })
+      } else {
+        toast({
+          title: 'Successfully deleted pet!',
+          description: 'Your pet is gone.',
+          status: 'success',
+          duration: 3000,
+          isClosable: true
+        })
+        setTimeout(function () {
+          router.reload()
+        }, 1000)
+      }
     }
   }
 
   return (
     <>
     <div className="bg-white w-[350px] rounded-lg shadow-lg">
-        <Image src="/download.jpeg" alt="this is a text" className="overflow-auto rounded-t-lg" width={350} height={100}/>
+        <Image src={pet.image} alt="this is a text" className="overflow-auto rounded-t-lg" width={350} height={100}/>
         <div className="py-5 px-5">
             <div><b>Name: </b>{pet.pet_name}</div>
             <div><b>Type: </b>{pet.type}</div>
@@ -71,7 +70,7 @@ export default function PetCard({pet}) {
                 />
             </button>
             
-            <button>
+            <button onClick={() => handleDelete(pet.pet_id)}>
                 <Image
                 src={"/delete.svg"}
                 alt="delete"
